fix(relay): reap idle half-closed TCP relays to avoid CLOSE_WAIT leak
When a peer half-closed (sent FIN) while the other side kept its half of
the stream open and idle, the relay's bidirectional copy waited forever
for the second EOF. The still-open socket then lingered in CLOSE_WAIT for
the entire lifetime of the parent QUIC connection, accumulating until the
process was restarted.

Add a shared `tuic_core::io::copy_bidirectional` with a half-close idle
reaper: it arms only after one direction reaches EOF, resets on any byte
transferred (so slow-but-active transfers survive), and never reaps a
fully-open idle tunnel. A zero timeout disables it.

- server: replace the hand-rolled `copy_io`, which had the same
  wait-for-both-EOF flaw, and reap after the configurable `stream_timeout`
  (60s default)
- client: socks5 inbound and TCP forward relays reap after 30s
