+++ tuic-core/Cargo.toml
@@ -33,14 +33,14 @@ serde = { version = "1", default-features = false, features = ["derive", "std"] 
 tracing = { version = "0.1", default-features = false}
 quinn = { workspace = true, default-features = false, features = ["futures-io", "runtime-tokio"]}
 quinn-congestions = { workspace = true }
-tokio = { version = "1", default-features = false, features = ["io-util"] }
+tokio = { version = "1", default-features = false, features = ["io-util", "macros", "time"] }
 eyre = { version = "0.6" }
 
 [dev-dependencies]
 tuic-core = { path = ".", features = ["async_marshal", "marshal", "model"] }
 uuid = { version = "1", features = ["v4"] }
 serde_json = { version = "1" }
-tokio = { version = "1", default-features = false, features = ["macros", "rt", "net", "io-util", "time"] }
+tokio = { version = "1", default-features = false, features = ["macros", "rt", "net", "io-util", "time", "test-util"] }
 rustls = { version = "0.23", default-features = false, features = ["std", "tls12", "ring"] }
 rcgen = { version = "0.14", default-features = false, features = ["crypto", "ring"] }
 tokio-rustls = { version = "0.26", default-features = false, features = ["ring"] }
